code refactoring and organisation into directories
