+++ .gitignore
@@ -1,0 +1,5 @@
+# Ignore all .exe files
+*.exe
+
+# Ignore Visual Studio Code files and folders
+.vscode/
