+++ Container.cpp
@@ -1,0 +1,8 @@
+#include <iostream>
+#include <string>
+
+#include "Container.h"
+
+Container::Container(const std::string &id) : id(id) {}
+
+Container::~Container() {}
