+++ ContainerStack.h
@@ -1,10 +1,7 @@
 #ifndef CONTAINERSTACK_H 
 #define CONTAINERSTACK_H
 
-#include<memory>
-#include <vector>
-
-#include "containers.h"
+#include "Container.h"
 
 class ContainerStack {
 public:
