+++ InitialStateReader.cpp
@@ -1,4 +1,6 @@
 #include <fstream>
+#include <string>
+#include <vector>
 
 #include "InitialStateReader.h"
